[REFACTOR] Move marketplace resource constants into evy-types
Web tests were importing MARKETPLACE_SERVICE and MARKETPLACE_RESOURCE
directly from services/marketplace/src/resources.ts, crossing a package
boundary. When web's tsc followed that import it tried to resolve
'evy-types' from within services/marketplace/src/resources.ts, which
fails in CI because setup-bun only installs web's node_modules (not
services/marketplace's), so services/marketplace/node_modules/evy-types
does not exist.

Move the identity constants (MARKETPLACE_SERVICE, MARKETPLACE_RESOURCE,
MARKETPLACE_RESOURCE_MANIFEST, MARKETPLACE_SEED_RESOURCES) into a new
evy-types/marketplaceResources module. The marketplace service re-exports
them from there (keeping its public API unchanged for data.ts/rpc.ts),
and the three web test files now import from evy-types/marketplaceResources.

Verified: web typecheck passes, marketplace build + lint + unit tests pass,
affected web unit tests pass (59/59).
